+++ snips_nlu_metrics/utils/dataset_utils.py
@@ -1,14 +1,13 @@
-from __future__ import absolute_import
-from __future__ import unicode_literals
+from __future__ import absolute_import, unicode_literals
 
 from collections import defaultdict
 from copy import deepcopy
 
-from future.utils import iteritems
+from future.utils import iteritems, itervalues
 
 from snips_nlu_metrics.utils.constants import (
-    ENTITIES, DATA, VALUE, USE_SYNONYMS, SYNONYMS, INTENTS, UTTERANCES, ENTITY,
-    TEXT)
+    DATA, ENTITIES, ENTITY, INTENTS, SYNONYMS, TEXT, USE_SYNONYMS, UTTERANCES,
+    VALUE)
 
 
 def input_string_from_chunks(chunks):
@@ -23,7 +22,7 @@ def get_utterances_subset(utterances, ratio):
         utterances_dict[intent_name].append(deepcopy(utterance))
 
     utterances_subset = []
-    for (intent_name, utterances) in utterances_dict.items():
+    for (intent_name, utterances) in iteritems(utterances_dict):
         nb_utterances = int(ratio * len(utterances))
         utterances_subset += [(intent_name, u)
                               for u in utterances[:nb_utterances]]
@@ -51,7 +50,7 @@ def get_declared_entities_values(dataset):
 
 def get_intent_utterances_entities_value(dataset):
     existing_entities = defaultdict(set)
-    for intent in dataset[INTENTS].values():
+    for intent in itervalues(dataset[INTENTS]):
         for u in intent[UTTERANCES]:
             for chunk in u[DATA]:
                 if ENTITY not in chunk or is_builtin_entity(chunk[ENTITY]):